refactor(CosiApplicationTests): 업비트 HTTP 클라이언트 테스트 리팩터링
CosiApplicationTests 위치 변경
test.java.com.cosi.api. -> test.java.com.cosi.

업비트 HTTP 클라이언트 테스트 로직 분리
test.java.com.cosi.upbit.UpbitHttpClientTests

테스트용 gson 속성으로 분리

테스트 메서드 이름 변경
- UpbitClientConfig에_등록된_RestClient가_잘_동작하는_지_눈으로_확인 → checkRestClient
- UpbitHttpClient_빈으로_종목_리스트_요청_테스트 → checkUpbitHttpClient
- UpbitMarkets_빈이_종목_리스트를_잘_업데이트_하는_지_테스트 → checkUpbitMarkets

관련 이슈
#1
