+++ api/src/test/java/com/cosi/CosiApplicationTests.java
@@ -1,0 +1,12 @@
+package com.cosi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CosiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+}
